немало

+++ services/media-service/src/main/java/org/ultra/rcrs/mediaservice/dao/model/Audio.java
@@ -30,7 +30,7 @@ public class Audio {
     private String container;
 
     @Column(name = "duration_ms")
-    private Long durationMs;
+    private Integer durationMs;
 
     private String bitrate;
